% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvinereg.R
\name{cvinereg}
\alias{cvinereg}
\title{C-vine regression models}
\usage{
cvinereg(
  formula,
  data,
  family_set = "parametric",
  selcrit = "aic",
  order = NA,
  par_1d = list(),
  weights = numeric(),
  cores = 1,
  ...,
  uscale = FALSE
)
}
\arguments{
\item{formula}{an object of class "formula"; same as \code{\link[=lm]{lm()}}.}

\item{data}{data frame (or object coercible by \code{\link[=as.data.frame]{as.data.frame()}}) containing
the variables in the model.}

\item{family_set}{see \code{family_set} argument of \code{\link[rvinecopulib:bicop]{rvinecopulib::bicop()}}.}

\item{selcrit}{selection criterion based on conditional log-likelihood.
\code{"loglik"} (default) imposes no correction; other choices are
\code{"aic"} and \code{"bic"}.}

\item{order}{the order of covariates in the C-vine, provided as vector of
variable names (after calling
\code{cvinereg:::expand_factors(model.frame(formula, data))}); selected
automatically if \code{order = NA} (default).}

\item{par_1d}{list of options passed to \code{\link[kde1d:kde1d]{kde1d::kde1d()}}, must be one value
for each margin, e.g. \code{list(xmin = c(0, 0, NaN))} if the response and first
covariate have non-negative support.}

\item{weights}{optional vector of weights for each observation.}

\item{cores}{integer; the number of cores to use for computations.}

\item{...}{further arguments passed to \code{\link[rvinecopulib:bicop]{rvinecopulib::bicop()}}.}

\item{uscale}{if TRUE, vinereg assumes that marginal distributions have been
taken care of in a preliminary step.}
}
\value{
An object of class cvinereg. It is a list containing the elements
\describe{
\item{formula}{the formula used for the fit.}
\item{selcrit}{criterion used for variable selection.}
\item{model_frame}{the data used to fit the regression model.}
\item{margins}{list of marginal models fitted by \code{\link[kde1d:kde1d]{kde1d::kde1d()}}.}
\item{vine}{an \code{\link[rvinecopulib:vinecop_dist]{rvinecopulib::vinecop_dist()}} object containing the fitted
C-vine.}
\item{stats}{fit statistics such as conditional
log-likelihood/AIC/BIC and p-values for each variable's contribution.}
\item{order}{order of the covariates chosen by the variable selection
algorithm.}
\item{selected_vars}{indices of selected variables.}}
Use \code{\link[=predict.cvinereg]{predict.cvinereg()}} to predict conditional quantiles. \code{summary.cvinereg()}
shows the contribution of each selected variable with the associated
p-value derived from a likelihood ratio test.
}
\description{
Sequential estimation of a regression C-vine for the purpose of quantile
prediction as described in Tepegjozova and Czado (2019) and Tepegjozova et al. (2022).
}
\details{
Note, that currently only continuous variables for the C-vine regression models are supported.
Furthermore, the code is heavily based on the one of the R-package \code{vinereg}, htpps://doi.org/10.32614/CRAN.package.vinereg.
}
\examples{
# simulate data
x <- matrix(rnorm(200), 100, 2)
y <- x \%*\% c(1, -2)
data <- data.frame(y = y, x = x)

# fit vine regression model
fit <- cvinereg(y ~ ., data)

# model predictions
mu_hat <- predict(fit, newdata = data, alpha = NA) # mean
med_hat <- predict(fit, newdata = data, alpha = 0.5) # median

}
\references{
Tepegjozova and Czado (2019), D- and C-vine quantile regression for large data sets,
https://mediatum.ub.tum.de/doc/1536070/1536070.pdf.

Tepegjozova et al. (2022). Nonparametric C- and D-vine-based quantile regression,
Dependence Modeling, 10(1), 1-21, https://doi.org/10.1515/demo-2022-0100.
}
\seealso{
\code{\link{predict.cvinereg}}
}
