% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_effects}
\alias{plot_effects}
\title{Plot marginal effects of a C-vine regression model}
\usage{
plot_effects(object, alpha = c(0.1, 0.5, 0.9), vars = object$order)
}
\arguments{
\item{object}{a \code{cvinereg} object.}

\item{alpha}{vector of quantile levels.}

\item{vars}{vector of variable names.}
}
\description{
The marginal effects of a variable is the expected effect, where expectation
is meant with respect to all other variables.
}
\examples{
# simulate data
x <- matrix(rnorm(200), 100, 2)
y <- x \%*\% c(1, -2)
data <- data.frame(y = y, x = x)

# fit vine regression model
fit <- cvinereg(y ~ ., data)

# plot
plot_effects(fit)

}
